requirements.txt
openai>=1.30.0
fastapi>=0.115.0
uvicorn[standard]
pandas
httpx
beautifulsoup4
pymupdf
chromadb
sentence-transformers
pydantic>=2.0
python-dotenv
